izp/requirements.txt
django-easy-pdf >= 0.1.1
xhtml2pdf>=0.2b1
